+++ .nvmrc
@@ -1,1 +1,1 @@
-v12.18.2
+v16.16.0
