--- lib/common/index.ts
@@ -1,3 +1,0 @@
-export * from './twilio.constants';
-export * from './twilio.decorator';
-export * from './twilio.utils';
