--- lib/common/twilio.constants.ts
@@ -1,2 +1,0 @@
-export const TWILIO_CONFIG_TOKEN = '__twilioCfg__';
-export const TWILIO_CONFIG_OPTIONS = '__twilioCfgOptions__';
