--- lib/interfaces/index.ts
@@ -1,3 +1,0 @@
-export * from './twilio-options.interface';
-export * from './twilio-decorator.interface';
-export * from './twilio-client.interface';
