--- lib/interfaces/twilio-client.interface.ts
@@ -1,4 +1,0 @@
-import * as Twilio from 'twilio';
-
-// eslint-disable-next-line
-export interface TwilioClient extends Twilio.Twilio {}
