--- lib/interfaces/twilio-decorator.interface.ts
@@ -1,7 +1,0 @@
-export interface TwilioDecorator {
-  (
-    target: Record<string, unknown>,
-    key: string | symbol,
-    index?: number | undefined,
-  ): void;
-}
