+++ lib/module/index.ts
@@ -1,0 +1,2 @@
+export * from './twilio.module';
+export * from './twilio.service';
