--- lib/providers/index.ts
@@ -1,1 +1,0 @@
-export * from './twilio.provider';
