+++ lib/utils/twilio.interface.ts
@@ -1,0 +1,9 @@
+import { Twilio } from 'twilio';
+
+export type TwilioClient = Twilio;
+
+export interface TwilioModuleOptions {
+  accountSid: string | undefined;
+  authToken: string | undefined;
+  options?: Twilio.TwilioClientOptions | undefined;
+}
