+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "nestjs-twilio",
-  "version": "2.2.1",
+  "version": "3.0.0",
   "description": "Injectable Twilio client for Nestjs",
   "license": "MIT",
   "author": {
     "name": "Lazar Karic",
-    "email": "kadzajr@gmail.com",
-    "url": "https://lkaric.tech"
+    "email": "lazar@mm.st",
+    "url": "https://lazar.sh"
   },
   "repository": "https://github.com/rejvban/nestjs-twilio",
   "bugs": "https://github.com/rejvban/nestjs-twilio/issues",
@@ -43,32 +43,32 @@
     "postpublish": "pinst --enable"
   },
   "devDependencies": {
-    "@nestjs/common": "8.4.4",
-    "@nestjs/core": "8.4.4",
-    "@nestjs/testing": "8.4.4",
-    "@types/jest": "27.4.1",
-    "@types/node": "17.0.23",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "dotenv": "^16.0.0",
-    "eslint": "^8.13.0",
+    "@nestjs/common": "9.0.8",
+    "@nestjs/core": "9.0.8",
+    "@nestjs/testing": "9.0.8",
+    "@types/jest": "28.1.6",
+    "@types/node": "18.6.4",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
+    "dotenv": "^16.0.1",
+    "eslint": "^8.21.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^7.0.4",
-    "jest": "27.5.1",
-    "lint-staged": "^12.3.7",
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.1",
+    "jest": "28.1.3",
+    "lint-staged": "^13.0.3",
     "pinst": "^3.0.0",
-    "prettier": "2.6.2",
+    "prettier": "2.7.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "7.5.5",
-    "ts-jest": "27.1.4",
-    "twilio": "3.76.1",
-    "typescript": "4.6.3"
+    "rxjs": "7.5.6",
+    "ts-jest": "28.0.7",
+    "twilio": "3.80.0",
+    "typescript": "4.7.4"
   },
   "peerDependencies": {
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
+    "@nestjs/common": "^9.0.0",
+    "@nestjs/core": "^9.0.0",
     "twilio": "^3.63.1"
   }
 }
